pyproject.toml
[project]
name = "klinedata48provider"
version = "0.1.0"
description = ""
authors = [
    {name = "wikimodels",email = "young.wiki.models@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi (==0.111.0)",
    "uvicorn[standard] (==0.29.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "aiohttp (==3.13.0)",
    "requests (==2.32.3)",
    "tenacity (==8.3.0)",
    "pandas (==2.2.2)",
    "numpy (==1.26.4)",
    "numba (==0.61.2)",
    "scipy (==1.13.1)",
    "hurst (==0.0.5)",
    "redis (==7.0.1)",
    "python-dotenv (==1.1.1)",
    "pytest (>=9.0.1,<10.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
